[DRAFT] some updates on broadcasting messages and testing

+++ shared/src/lib.rs
@@ -1,3 +1,6 @@
+use serde::{Serialize, Deserialize};
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
 pub enum ClientMessage {
     LinesDestroyed (
         // u16 since it's boolean anyway, we can use an int as an array of bits, don't need sign nor more than 16 wide
@@ -7,7 +10,8 @@ pub enum ClientMessage {
     GameOver,
 }
 
+#[derive(Serialize, Deserialize, Debug)]
 pub enum ServerMessage {
-    Broadcast { user_id: i32, msg: ClientMessage },
+    Broadcast { user_id: u128, msg: ClientMessage },
     LeaderBoardUpdate(),
 }
